refactored api version specification and requester
