+++ lib/blurb/account.rb
@@ -6,7 +6,7 @@ class Blurb
   class Account < BaseClass
     attr_accessor :refresh_token, :api_url, :client, :profiles, :active_profile
 
-    delegate :api_version, to: :client
+    delegate :reports_api_version, to: :client
 
     API_URLS = {
       "TEST" => "https://advertising-api-test.amazon.com",
