+++ lib/blurb/profile.rb
@@ -2,12 +2,12 @@ require "blurb/account"
 require "blurb/campaign_requests"
 require "blurb/snapshot_requests"
 require "blurb/report_requests"
+require "blurb/report_requests_v3"
 require "blurb/request_collection"
 require "blurb/request_collection_with_campaign_type"
 require "blurb/suggested_keyword_requests"
 require "blurb/history_request"
 require "blurb/invoice_request"
-require "blurb/v3/report_requests"
 
 class Blurb
   class Profile < BaseClass
@@ -93,8 +93,8 @@ class Blurb
         base_url: @account.api_url,
         campaign_type: :hsa
       )
-      @v3_reports = V3::ReportRequests.new(
-        headers: headers_hash(v3: true),
+      @v3_reports = ReportRequestsV3.new(
+        headers: headers_hash(reports_v3: true),
         base_url: @account.api_url
       )
       @ad_groups = RequestCollection.new(
@@ -174,7 +174,7 @@ class Blurb
     end
 
     def reports(campaign_type = nil)
-      return @v3_reports if @account.api_version == 3
+      return @v3_reports if @account.reports_api_version == 3
       return @sp_reports if campaign_type == :sp
       return @sb_reports if campaign_type == :sb || campaign_type == :hsa
       return @sd_reports if campaign_type == :sd
@@ -209,7 +209,7 @@ class Blurb
       }
 
       headers_hash["Content-Encoding"] = "gzip" if opts[:gzip]
-      headers_hash["Content-Type"] = 'application/vnd.createasyncreportrequest.v3+json' if opts[:v3]
+      headers_hash["Content-Type"] = 'application/vnd.createasyncreportrequest.v3+json' if opts[:reports_v3]
 
       return headers_hash
     end
